# snapshot: use io.Copy instead of copying parts row by row