Previously, we were reading rows one by one from parts and writing them to disk on snapshot. However, we already have a serialized version of the parquet bytes by accessing the parquet file. This commit uses io.Copy to copy those bytes directly to disk. This commit halves snapshot time.

```
$ benchstat benchmain benchcopy
goos: darwin
goarch: arm64
pkg: github.com/polarsignals/frostdb
            │  benchmain  │             benchcopy             │
            │   sec/op    │   sec/op    vs base               │
Snapshot-12   36.59 ± 19%   18.91 ± 2%  -48.31% (p=0.002 n=6)
```